changed all possition variables to doubles

+++ include/building.h
@@ -2,7 +2,7 @@
 #include "datatypes.h"
 #include <SDL2/SDL.h>
 
-void buildingCreate(GameHandler* gameHandler, int pos, int stories);
+void buildingCreate(GameHandler* gameHandler, double pos, int stories);
 void buildingRemove(GameHandler* gameHandler, int element);
 
 void buildingExitSoldier();
