+++ include/missilePad.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "datatypes.h"
 
-void missilePadCreate(GameHandler* gameHandler, int pos);
+void missilePadCreate(GameHandler* gameHandler, double pos);
 void missilePadRemove(GameHandler* gameHandler, int element);
 
 void missilePadLaunch(Vec2 helecopterDirection);
